requirements.txt
bcrypt==3.1.7
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
constantly==15.1.0
cryptography==2.7
cssselect==1.1.0
Django==2.2.4
django-cors-headers==3.1.0
django-debug-toolbar==2.0
httpie==1.0.3
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
lxml==4.4.1
mysql-connector-python==8.0.5
mysqlclient==1.4.4
parsel==1.5.2
protobuf==3.9.1
pyasn1==0.4.7
pyasn1-modules==0.2.6
pycparser==2.19
PyDispatcher==2.0.5
Pygments==2.4.2
PyHamcrest==1.9.0
PyJWT==1.7.1
pyOpenSSL==19.0.0
pytz==2019.2
queuelib==1.5.0
requests==2.22.0
Scrapy==1.7.3
service-identity==18.1.0
six==1.12.0
SQLAlchemy==1.3.7
sqlparse==0.3.0
style==1.1.0
Twisted==19.7.0
update==0.0.1
urllib3==1.25.3
w3lib==1.21.0
zope.interface==4.6.0
